bazel-vfs: hack-tastical start to prototype Git sparse-checkout integration
Currently commandeers the filesystem data structure to inject a 'git
sparse-checkout add'. Works with mixed success due to Git locking issues
(there's a *lot* of nonsense calls being made as a result of making this
happen over the entire filesystem).

Next step: restrict this behavior to only the PathPackageLocator looking for
things in source directories.

Signed-off-by: Victoria Dye <vdye@github.com>